M    mert/TERsrc/hashMap.cpp M    mert/TERsrc/hashMap.h
git-svn-id: https://mosesdecoder.svn.sourceforge.net/svnroot/mosesdecoder/branches/mert-other_metrics@3817 1f5c12ca-751b-0410-a591-d2e778427230

+++ mert/TERsrc/hashMap.h
@@ -23,7 +23,7 @@ namespace HashMapSpace
         public:
 //     ~hashMap();
             long hashValue ( string key );
-            int trouve ( size_t searchKey );
+            int trouve ( long searchKey );
             int trouve ( string key );
             void addHasher ( string key, string value );
             stringHasher getHasher ( string key );
